getdist: removed argument "prior_mrpredictive=None" from PosteriorAnalys.mass_radius_prior_predictive_plot()
This argument could be used to load an MR_prpr.txt from disk and plot it, but it is not used anywhere in the NEoST code. I think it was meant to be used for comparing two M-R plots, because another M-R plot is drawn regardless of whether the argument is supplied. I removed this functionality because it relies on outdated getdist levels. A user can write such a comparison function themselves if they want to.
	modified:   neost/PosteriorAnalysis.py