seaborn kdeplot and getdist: updated "levels", "fill", and "cmap" arguments to kdeplot.
Removed the old "n_levels" argument to kdeplot (which took levels from getdist) and replaced it with "levels" which accepts a simple list of percentages. This allows us to remove getdist as a dependency completely. (Thus removed it from conf.py.)
The "shade" argument has been renamed "fill" in seaborn, so I changed it.
The "colors" argument no longer works in kdeplot. I instead feed a ListedColormap to the "cmap" argument, achieving the same look.

	modified:   docs/source/conf.py
	modified:   neost/PosteriorAnalysis.py