numpy: avoided ragged arrays in Python TOV solver
The python TOV solver sometimes created ragged arrays, which are unsupported in numpy >= 1.24. Some values in the initial_conditions() function were sometimes scalars and sometimes arrays of length 1. I introduced checks to make sure that if these values are arrays, the length is indeed 1. If so, convert the array to a scalar. If the conditions are unmet then a ValueError is raised.

	modified:   neost/tovsolvers/TOVr_python.py